refaktor GeoRep
